refactor: extract shared get_line_range util
Move get_line_range from process_diagnostics to a
shared utility. Add column data to diagnostics.
